requirements.txt
absl-py==1.4.0
aiocache==0.11.1
aiocqhttp==1.4.2
aiodns==3.0.0
aiofiles==0.7.0
aiohttp==3.8.3
aiosignal==1.2.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
altgraph==0.17.2
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.8.1
arrow==1.2.3
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.25.0
atomicwrites==1.4.0
attrs==21.2.0
audioread==3.0.0
backports.zoneinfo==0.2.1
bbcode==1.1.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
blinker==1.4
Brotli==1.0.9
brotlipy==0.7.0
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.3.0
cattrs==1.8.0
cchardet==2.1.7
certifi==2022.12.7
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==5.1.0
charset-normalizer==2.1.1
cheroot==8.5.2
cleo==2.0.1
click==8.1.3
colorama==0.4.6
cookiecutter==1.7.3
crashtest==0.4.1
crcmod==1.7
cryptography @ file:///C:/b/abs_36x9ifdcl4/croot/cryptography_1665612655344/work
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.33
decorator==4.4.2
distlib==0.3.6
dulwich==0.20.50
emoji==1.7.0
fastapi==0.79.1
feedparser==6.0.8
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.9.0
filetype==1.0.10
flatbuffers==23.3.3
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.28.2
frozenlist==1.2.0
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
gevent==21.8.0
gino==1.0.1
gocqapi==0.1.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.3
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hjson==3.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
httpx-socks==0.7.5
hypercorn==0.12.0
hyperframe==6.0.1
idna==3.4
imageio==2.13.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.13.0
importlib-resources==5.10.2
iniconfig==1.1.1
itsdangerous==2.0.1
jamo==0.4.1
jaraco.classes==3.2.3
jaraco.functools==3.4.0
jieba==0.42.1
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.17.3
keras==2.11.0
keyring==23.13.1
kiwisolver==1.3.2
langid==1.1.6
libclang==15.0.6.1
librosa==0.8.1
llvmlite==0.39.1
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.0
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
more-itertools==8.11.0
moviepy==1.0.3
msgpack==1.0.4
multidict==6.0.4
mutagen==1.45.1
nb-cli==0.6.9
nest-asyncio==1.5.6
nonebot-adapter-cqhttp==2.0.0a16
nonebot-adapter-onebot==2.1.5
nonebot-plugin-apscheduler==0.1.2
nonebot-plugin-guild-patch==0.2.2
nonebot-plugin-hanayori==0.5
nonebot-plugin-imageutils==0.1.13.4
nonebot-plugin-picsearcher==0.1.6rc2
nonebot-plugin-RealESRGAN==0.2.1
nonebot-plugin-reboot==0.1.2
nonebot-plugin-setu==0.0.6
nonebot-plugin-status==0.4.0
nonebot-plugin-tts-gal==0.3.6
nonebot2==2.0.0b5
norbert==0.2.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
oss2==2.15.0
outcome==1.1.0
packaging==21.3
pandas==1.5.3
pefile==2021.9.3
peony-twitter==1.1.7
pexpect==4.8.0
picimagesearch==3.7.8
Pillow==9.3.0
pkginfo==1.9.6
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
playwright==1.29.1
pluggy==1.0.0
poetry==1.3.1
poetry-core==1.4.0
poetry-plugin-export==1.2.0
pooch==1.7.0
poyo==0.5.0
priority==2.0.0
proglog==0.1.9
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
py==1.11.0
pyahocorasick==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.15.0
pydantic==1.9.2
pydub==0.25.1
pyee==9.0.4
pyfiglet==0.8.post1
pygtrie==2.5.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
pymdown-extensions==9.9
pyopenjtalk==0.3.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==2.4.7
pypinyin==0.47.1
pyquery==2.0.0
pyrsistent==0.19.3
PySocks @ file:///C:/ci/pysocks_1605287845585/work
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-magic==0.4.24
python-markdown-math==0.8
python-slugify==7.0.0
python-socks==2.0.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
Quart==0.14.1
rapidfuzz==2.13.7
requests==2.28.1
requests-cache==0.8.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
resampy==0.4.2
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scheduler==0.6.1
scikit-learn==1.2.1
scipy==1.10.0
selenium==2.48.0
setuptools-scm==6.3.2
sgmllib3k==1.0.0
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sogou-tr-free==0.0.11
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.2.1
SQLAlchemy==1.3.24
starlette==0.19.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinydb==4.5.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
trove-classifiers==2022.12.22
typeguard==2.13.2
typer==0.3.2
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.1
ujson==4.0.2
Unidecode==1.3.6
url-normalize==1.4.3
urllib3==1.26.13
uvicorn==0.18.3
virtualenv==20.17.1
watchfiles==0.18.1
watchgod==0.7
wcwidth==0.2.5
web.py==0.62
webencodings==0.5.1
webrtcvad-wheels==2.0.11.post1
websocket==0.2.1
websockets==10.4
Werkzeug==2.0.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
win32-setctime==1.1.0
wrapt==1.15.0
wsproto==1.0.0
yarl==1.8.2
zhconv==1.4.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
